requirements.txt
streamlit
sentence-transformers
scikit-learn
PyPDF2
google-generativeai
pandas
numpy